Massive refactoring: move a lot of helpers to `utils.js`.
Fix several tiny bugs (detected by eslint)
