+++ lib/client.js
@@ -5,75 +5,9 @@
 
 "use strict";
 
-const request = require('request');
-const debug = require('debug')('freshdesk-api');
 
-
-const createResponseHandler = function (cb) {
-	return function (error, response, body) {
-		if (error) {
-			debug(`Unexpected status [${error}], req path [${response.request.path}] raw body: ${response.request.body}`);
-			return cb(error);
-		}
-
-		let data = null;
-
-		switch(response.statusCode) {
-		// SUCCESS
-		// https://httpstatuses.com/200 OK
-		// https://httpstatuses.com/201 Created
-		case 200:
-		case 201:
-			try {
-				data = JSON.parse(body);
-			} catch (err) {
-				// TODO: catch only JSON-parse errors!!!
-				debug(`Not a JSON resp with valid status [${response.statusCode}], req path [${response.request.path}] raw body: ${response.request.body}`);
-				return cb(new FreshdeskError('Not a JSON response from API', data));
-			}
-			return cb(null, data);
-
-		// SUCCESS for DELETE operations
-		// https://httpstatuses.com/204 No Content
-		case 204:
-			return cb(null, null);
-
-		// https://httpstatuses.com/409 Conflict  - NOT UNIQUE, where unique required
-		case 409:
-		default:
-			try {
-				data = JSON.parse(body);
-			} catch (err) {
-				// TODO: catch only JSON-parse errors!!!
-
-				// probably, data is not a JSON, so lets return it "AS IS"
-				debug(`Not a JSON resp for unexpected status [${response.statusCode}], req path [${response.request.path}] raw body: ${response.request.body}`);
-				return cb(new FreshdeskError('Not a JSON response from API', data));
-			}
-
-			debug(`Unexpected/Error status [${response.statusCode}], req path [${response.request.path}] raw body: ${response.request.body}`);
-			return cb(new FreshdeskError(data.description, data));
-		}
-	}
-}
-
-const makeRequest = function (method, auth, url, qs, data, cb) {
-	var options = {
-		method: method,
-		headers: {
-			'Content-Type': 'application/json',
-			'Authorization': auth
-		},
-		url: url, // for debugging set to: "https://httpbin.org/get"
-		qs: qs
-	}
-
-	if(data) {
-		options.body = JSON.stringify(data);
-	}
-
-	request(options, createResponseHandler(cb))
-}
+const makeRequest     = require('./utils').makeRequest;
+const FreshdeskError  = require('./utils').FreshdeskError;
 
 
 /**
@@ -305,26 +239,12 @@ Freshdesk.prototype = Object.create({
 	 * @param  {integer}                            id      Company ID
 	 * @param  {Freshdesk.requestCallback}          cb      Callback function {@link Freshdesk.requestCallback}
 	 */
-	deleteCompany(id, data, cb) {
+	deleteCompany(id, cb) {
 		makeRequest('DELETE', this._auth, `${this.baseUrl}/api/v2/companies/${id}`, null, null, cb);
 	},
 
 });
 
-/**
- * Freshdesk's API protocol violations
- *
- * @param {[type]} message [description]
- * @param {[type]} data    [description]
- */
-function FreshdeskError(message, data) {
-  this.name = 'FreshdeskError';
-  this.message = message || 'Error in Freshdesk\'s client API';
-  this.stack = (new Error()).stack;
-  this.data = data;
-}
-FreshdeskError.prototype = Object.create(Error.prototype);
-FreshdeskError.prototype.constructor = FreshdeskError;
 
 /* EXPORT SECTION */
 module.exports = Freshdesk;
